refactor(BookCatalog): move into a new module
